backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyMuPDF
pandas
openpyxl
